Including New Setup That Should Work for Gurobi 10.0.0
